Staging2.0

+++ Frontend/src/views/app/Dashboard.js
@@ -1,30 +1,29 @@
 import React, { useState, useEffect, Fragment } from 'react';
+import * as urls from '../../Urls'
 
 export const Dashboard = () => {
   const [userName, setUserName] = useState('');
   const [loading, setLoading] = useState(true);
 
   useEffect(() => {
-    console.log('before if dashboard')
     if (!localStorage.getItem('token')) {
-      console.log('inside first condition')
-      window.location.replace('http://localhost:8080/login');
+      window.location.replace(`${urls.origin}/login`);
     } else {
-      console.log('inside second condition')
-      fetch('http://localhost:8000/users/auth/user/', {
+      fetch(`${urls.api}/users/auth/user/`, {
         method: 'GET',
         headers: {
           'Content-Type': 'application/json',
           Authorization: `Token ${localStorage.getItem('token')}`
         }
       })
-        .then(res => {if(res.status === 200){
-                        return res.json()
-                      }else{
-                        localStorage.clear()
-                        window.location.replace('http://localhost:8080/login');
-                      }
-                    })
+        .then(res => {
+          if (res.status === 200) {
+            return res.json()
+          } else {
+            localStorage.clear()
+            window.location.replace(`${urls.origin}/login`);
+          }
+        })
         .then(data => {
           setUserName(data.username);
           setLoading(false);
@@ -34,7 +33,7 @@ export const Dashboard = () => {
 
   return (
     <div>
-      {loading === false && (  
+      {loading === false && (
         <Fragment>
           <h1>Dashboard</h1>
           <h2>Hello {userName}!</h2>
